+++ newton/_src/solvers/kamino/_src/linalg/linear.py
@@ -776,7 +776,7 @@ class ConjugateGradientSolver(IterativeSolver[ScalarType, IndexType]):
             maxiter=self._maxiter,
             Mi=self._Mi,
             callback=None,
-            use_cuda_graph=True,
+            use_graph=True,
             use_graph_conditionals=self._use_graph_conditionals,
             loop_granularity=self.loop_granularity,
         )
@@ -790,7 +790,7 @@ class ConjugateGradientSolver(IterativeSolver[ScalarType, IndexType]):
                 maxiter=self._maxiter,
                 Mi=self._Mi,
                 callback=None,
-                use_cuda_graph=True,
+                use_graph=True,
                 use_graph_conditionals=self._use_graph_conditionals,
                 loop_granularity=self.loop_granularity,
             )
@@ -888,7 +888,7 @@ class ConjugateResidualSolver(IterativeSolver[ScalarType, IndexType]):
             maxiter=self._maxiter,
             Mi=self._Mi,
             callback=None,
-            use_cuda_graph=True,
+            use_graph=True,
             use_graph_conditionals=self._use_graph_conditionals,
             loop_granularity=self.loop_granularity,
         )
@@ -902,7 +902,7 @@ class ConjugateResidualSolver(IterativeSolver[ScalarType, IndexType]):
                 maxiter=self._maxiter,
                 Mi=self._Mi,
                 callback=None,
-                use_cuda_graph=True,
+                use_graph=True,
                 use_graph_conditionals=self._use_graph_conditionals,
                 loop_granularity=self.loop_granularity,
             )