+++ newton/examples/cable/example_cable_y_junction.py
@@ -148,12 +148,9 @@ class Example:
         self.capture()
 
     def capture(self):
-        if self.solver.device.is_cuda:
-            with wp.ScopedCapture() as capture:
-                self.simulate()
-            self.graph = capture.graph
-        else:
-            self.graph = None
+        with wp.ScopedCapture() as capture:
+            self.simulate()
+        self.graph = capture.graph
 
     def simulate(self):
         for _ in range(self.sim_substeps):