+++ newton/examples/diffsim/example_diffsim_drone.py
@@ -467,7 +467,7 @@ class Example:
         # We start with -1 since it'll be incremented on the first frame.
         self.target_idx = -1
         # use a Warp array to store the current target so that we can assign
-        # a new target to it while retaining the original CUDA graph.
+        # a new target to it while retaining the original graph.
         self.current_target = wp.array([self.targets[self.target_idx + 1]], dtype=wp.vec3)
 
         # Number of steps to run at each frame for the optimisation pass.